x86_64: Conditionally define __sfp_handle_exceptions for compiler-rt
The LLVM compiler-rt builtins library does not currently provide an
implementation for __sfp_handle_exceptions.  On x86_64, this causes
unresolved symbol errors when building glibc in environments that
exclude libgcc.

This patch implements __sfp_handle_exceptions specifically for x86_64,
bridging the gap for non-GNU compiler runtimes.

The implementation is used conditionally, only if the compiler does
not already provide the symbol.

NB: the implementation is based on libgcc and raises bosh SSE and i387
    exceptions (different that the one from 460ee50de054396cc9791ff4)