x86: Check if compiler runtime supports __sfp_handle_exceptions
And provides an implementation based on libgcc if compiler does not
support it.  The __sfp_handle_exceptions is used internally by x86
for the float128 support and it is usually implemented by libgcc.

The compiler-rt does not provide __sfp_handle_exceptions, so if
glibc is built --rtlib=compiler the libm.so fails with an undefined
symbol.