Repurpose have-cc-with-libunwind to which unwind library to use
The have-cc-with-libunwind/HAVE_CC_WITH_LIBUNWIND was used solely
on ia64 to use libunwind along with libgcc for unwind support.  Since
it is unused on gcc, this now checks which unwind compiler supports
to allow clang build use libunwind instead of libgcc_s.

For have-cc-with-libunwind -lclang_rt.builtins is added to libgcc_eh
because links-dso-program/links-dso-program-c fails to link due a
missing __gcc_personality_v0 (clang seems to not add it if rlink-path
is used).

TODO: check a better way to define -lclang_rt.builtins