nptl: Only initialize robust list at mutex usage
The robust mutex kernel initialization is always performed at process
startup and thread creation, even when the process does not use any
robust mutexes.  This patch moves the set_robust_issue syscall to
either when a robust process is initialized or used, saving the
syscall issue for processes that do not use robust mutexes.

The field futex_offset is used as a sentinel to indicate whether the
struct pthread list is already initialized.  This also allows to move
the __nptl_set_robust_list_avail from ld.so to libc.so, since there is
no need to initialize it at pthread startup.

Checked on x86_64-linux-gnu and aarch64-linux-gnu.