string: Add fallback implementation for ctz/clz
The algorithm is based on De Bruijn sequence, and it might be used
if the ABI does not want to use libgcc fallback implementation (for
instance on statup code where libgcc might be compiled with some
options not support at the time, like stack protector).

It is disable by default, so no target is affected.