Use _dl_writev on __libc_message_impl
And change _dl_writev to return an negative errno in case of failure.
This keeps the required semantic on not setting errno in case of
failure and allow remove the Linux libc_fatal.c implementation.

It also makes is simple to use writev syscall during process startup,
specially on i386 where it requires to disable the vDSO usage.