x86: Use -mavx instead of -msse2avx
clang supports -msse2avx from version 19 and onwards, but it should
be gated as an option to assembler (either with -Wa or -Xassembler).

The -DSSE2AVX option was used because there were asm statements with
SSE-only instructions which was fixed by commit ff8be6152bc.

Now we can simply use -mavx.

Reviewed-by: H.J. Lu <hjl.tools@gmail.com>