dlfcn: Fix dlclose crash during C++ thread_local destructor (BZ 33598)
If a dynamically loaded shared library uses thread_local and unloads
another library in its C++ destructor, glibc will crash because
dlclose(), which is called by the library, unmaps both libraries, and
control from dlclose() is returned to an unmapped page.

For instance:

  $ cat <<EOF > main.c
  #include <dlfcn.h>
  #include <assert.h>
  int main()
  {
    void *h = dlopen ("libt1.so", RTLD_NOW);
    assert (h);
    dlclose (h);
  }
  EOF
  $ cat <<EOF > t1.cpp
  #include <dlfcn.h>
  #include <assert.h>
  #include <memory>
  thread_local std::unique_ptr<int> tlp;
  static struct c1 {
    void *h;
    c1() {
      h = dlopen("libm.so", RTLD_NOW);
      assert(h);
      tlp = std::make_unique<int>();
    }
     ~c1() { dlclose(h); }
  } c1;
  EOF
  $ g++ -Wall -g -fPIC -shared -o libt1.so t1.cpp -ldl
  $ gcc -Wall -g -fPIC    main.c  -ldl -o main
  $ LD_LIBRARY_PATH=. ./main
  Segmentation fault

The issue is that during libt1.so dlclose, no module is unmap because
they are still in use (elf/dl-close.c:178). However, because the
libraries also create a thread-local variable in the main thread, the
struct destructor (c1::~c1) is called during __run_exit_handlers after
main returns.

The most straightforward solution, which does not require any change
when the object is unmapped or add any extra complexity to dependency
tracking, is to assume that all libraries are in use at exit (similar
to the RTLD_NODELETE semantics).

This is accomplished with a new rtld_global variable, dl_at_exit, set
only in __run_exit_handlers.  To avoid potential issues with malloc
debug, where __libc_freeres might not free some link_map data structures,
__libc_dlclose does not consider the variable value (which is
 implemented via an extra flag).

Checked on aarch64-linux-gnu, x86_64-linux-gnu, and i686-linux-gnu.