configure: replace -Ttext-segment -with --imgage-base
Although they are not strictly identical, the -Ttext-segment does not
consider the ELF header when setting the base virtual address; both
are sufficient to trigger the issue for the tests that use it [1]
(a non-zero load address).

Also, lld does not support -Ttext-segment.  The patch also changes the
onfigure test to use LIBC_TEST_LINKER_FEATURE, since it is used only
for testing.

[1] https://sourceware.org/bugzilla/show_bug.cgi?id=31799