elf: Allow vDSO as a direct dependency (BZ 33335)
The vDSO is treated as a special link map, and using it as a direct
dependency shows two issues:

1. Binding uses the implicit vDSO provided by the kernel (even though a
   library is explicitly provided);
2. And depending on how the binary is linked, it can cause loops that
   trigger an inconsistency during process startup.

Both issues can be shown with the following example:

  $ cat > repro.c << EOF
  #include <time.h>
  extern int __vdso_clock_gettime (clockid_t clk_id, struct timespec *tp);
  int main ()
  {
    __vdso_clock_gettime (CLOCK_REALTIME, &(struct timespec) {0});
  }
  EOF
  $ cat > vdso.c << EOF
  #include <stdio.h>
  #include <time.h>
  int __vdso_clock_gettime (clockid_t clk_id, struct timespec *tp)
  {
    puts ( __func__);
  }
  EOF
  $ gcc -shared -fPIC vdso.c -o linux-vdso.so.1 -nodefaultlibs
  $ gcc repro.c -L. -l:linux-vdso.so.1 -o test1 && ./test1
  $ gcc repro.c -nodefaultlibs -L. -lc -l:linux-vdso.so.1 -o test2 && ./test2
  Inconsistency detected by ld.so: rtld.c: 2013: dl_main: Assertion `GL(dl_rtld_map).l_prev->l_next == GL(dl_rtld_map).l_next' failed!

The first issue, always binding to implicit vDSO, is caused by the vDSO
being present in the global link map list during _dl_lookup_map. The
second issue, triggered by test2, is trickier to understand, but also
due to the same behavior: after the program is loaded
(_dl_map_object_deps), the main_map search list contains the loader in a
position different than last one ({ "", "libc.so", "ld.so",
"linux-vdso.so.1" }). This causes a cycle in the search list's order,
triggering the assert.

To fix both issues, a new flag is added to link_map when the vDSO is
loaded (through __RTLD_VDSO), and the vDSO link map is ignored on
_dl_lookup_map. This allows programs to bind to a provided DSO, aiming
to mimic Linux vDSO.

The test requires specific rules to build a program that places the vDSO
at the end of the DT_NEEDED dynamic tags, and additional make rules to
handle different vDSO names across architectures.

Checked on x86_64-linux-gnu, i686-linux-gnu, aarch64-linux-gnu,
powerpc-linux-gnu, and powerpc64-linux-gnu.